Add files via upload
This code implements an object that can be used to locate the roots of 1D functions. The code searches a large interval for the sub-intervals known to contain roots. Improved approximations to the roots are determined using bisection method or Newton-Rhaphson method. This code could be extended by adding other root finding methods, but I've found that Newton-Rhaphson with derivative computation works quite well. If you didn't want the overhead of computing the derivative of a function you could include a technique such as Ridder's method.